requirements.txt
sqlalchemy
python-dateutil
ipython
slackclient
beautiful-soup